PIM-3882: improve steps
