+++ src/Pim/Bundle/EnrichBundle/Processor/MassEdit/AbstractMassEditProcessor.php
@@ -1,0 +1,67 @@
+<?php
+
+namespace Pim\Bundle\EnrichBundle\Processor\MassEdit;
+
+use Akeneo\Bundle\BatchBundle\Entity\StepExecution;
+use Akeneo\Bundle\BatchBundle\Item\AbstractConfigurableStepElement;
+use Akeneo\Bundle\BatchBundle\Step\StepExecutionAwareInterface;
+use Pim\Bundle\CatalogBundle\Model\ProductInterface;
+use Symfony\Component\Validator\ConstraintViolationListInterface;
+
+/**
+ * AbstractMassEditProcessor
+ *
+ * @author    Olivier Soulet <olivier.soulet@akeneo.com>
+ * @copyright 2015 Akeneo SAS (http://www.akeneo.com)
+ * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
+ */
+class AbstractMassEditProcessor extends AbstractConfigurableStepElement implements StepExecutionAwareInterface
+{
+    /** @var StepExecution */
+    protected $stepExecution;
+
+    /**
+     * {@inheritdoc}
+     */
+    public function setStepExecution(StepExecution $stepExecution)
+    {
+        $this->stepExecution = $stepExecution;
+
+        return $this;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getConfigurationFields()
+    {
+        return [];
+    }
+
+    /**
+     * @param ConstraintViolationListInterface $violations
+     * @param ProductInterface                 $product
+     */
+    protected function addWarningMessage(
+        ConstraintViolationListInterface $violations,
+        ProductInterface $product
+    ) {
+        foreach ($violations as $violation) {
+            // TODO re-format the message, property path doesn't exist for class constraint
+            // for instance cf VariantGroupAxis
+            $invalidValue = $violation->getInvalidValue();
+            if (is_object($invalidValue) && method_exists($invalidValue, '__toString')) {
+                $invalidValue = (string) $invalidValue;
+            } elseif (is_object($invalidValue)) {
+                $invalidValue = get_class($invalidValue);
+            }
+            $errors = sprintf(
+                "%s: %s: %s\n",
+                $violation->getPropertyPath(),
+                $violation->getMessage(),
+                $invalidValue
+            );
+            $this->stepExecution->addWarning($this->getName(), $errors, [], $product);
+        }
+    }
+}
