+++ src/Pim/Bundle/EnrichBundle/Processor/MassEdit/AddProductValueProcessor.php
@@ -2,14 +2,9 @@
 
 namespace Pim\Bundle\EnrichBundle\Processor\MassEdit;
 
-use Akeneo\Bundle\BatchBundle\Entity\StepExecution;
-use Akeneo\Bundle\BatchBundle\Item\AbstractConfigurableStepElement;
 use Akeneo\Bundle\BatchBundle\Item\ItemProcessorInterface;
-use Akeneo\Bundle\BatchBundle\Step\StepExecutionAwareInterface;
 use Pim\Bundle\CatalogBundle\Model\ProductInterface;
 use Pim\Bundle\CatalogBundle\Updater\ProductUpdaterInterface;
-use Pim\Bundle\EnrichBundle\MassEditAction\Handler\UpdateProductHandler;
-use Symfony\Component\Validator\ConstraintViolationListInterface;
 use Symfony\Component\Validator\ValidatorInterface;
 
 /**
@@ -19,13 +14,10 @@ use Symfony\Component\Validator\ValidatorInterface;
  * @copyright 2015 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class AddProductValueProcessor extends AbstractConfigurableStepElement implements ItemProcessorInterface, StepExecutionAwareInterface
+class AddProductValueProcessor extends AbstractMassEditProcessor implements ItemProcessorInterface
 {
     /** @var ProductUpdaterInterface */
     protected $productUpdater;
-
-    /** @var StepExecution */
-    protected $stepExecution;
 
     /** @var ValidatorInterface */
     protected $validator;
@@ -63,24 +55,6 @@ class AddProductValueProcessor extends AbstractConfigurableStepElement implement
     }
 
     /**
-     * {@inheritdoc}
-     */
-    public function setStepExecution(StepExecution $stepExecution)
-    {
-        $this->stepExecution = $stepExecution;
-
-        return $this;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getConfigurationFields()
-    {
-        return [];
-    }
-
-    /**
      * Add data from $actions to the given $product
      *
      * @param ProductInterface $product
@@ -95,32 +69,5 @@ class AddProductValueProcessor extends AbstractConfigurableStepElement implement
         }
 
         return $this;
-    }
-
-    /**
-     * @param ConstraintViolationListInterface $violations
-     * @param ProductInterface                 $product
-     */
-    protected function addWarningMessage(
-        ConstraintViolationListInterface $violations,
-        ProductInterface $product
-    ) {
-        foreach ($violations as $violation) {
-            // TODO re-format the message, property path doesn't exist for class constraint
-            // for instance cf VariantGroupAxis
-            $invalidValue = $violation->getInvalidValue();
-            if (is_object($invalidValue) && method_exists($invalidValue, '__toString')) {
-                $invalidValue = (string) $invalidValue;
-            } elseif (is_object($invalidValue)) {
-                $invalidValue = get_class($invalidValue);
-            }
-            $errors = sprintf(
-                "%s: %s: %s\n",
-                $violation->getPropertyPath(),
-                $violation->getMessage(),
-                $invalidValue
-            );
-            $this->stepExecution->addWarning($this->getName(), $errors, [], $product);
-        }
     }
 }
